command creator, ping reply and no pings ??? WTF IS THIS i need sleep

+++ function/sayCommand.js
@@ -1,9 +1,19 @@
 module.exports =
 
 function sayCommand(arguments, msg) {
+    if (msg.content.includes("@everyone") || msg.content.includes("@here")) {
+      msg.channel.send("Ora ora ora ora! Remove that massive ping.");
+      return
+    }
+
     const sayMessage = arguments.join(" ");
-    var checkPings = sayMessage.toString();
-    var finalMessage = checkPings.replace("@everyone", "everyone").replace("@here", "here");
+
+    if (sayMessage.includes("<@!") && msg.content.includes(">")) {
+      msg.channel.send("Avoid pings. What? Isn't that a ping?");
+      return
+    }
+
+    var finalMessage = sayMessage.toString();
     msg.delete().catch(O_o=>{});
     msg.channel.send(finalMessage);
 }
